create DbSets

+++ EntityFramework/Entities/Category.cs
@@ -6,15 +6,15 @@ using System.Threading.Tasks;
 
 namespace EntityFramework.Entities
 {
-   public class Category
+    public class Category
     {
-        
-            public int Id { get; set; }
-            public string Name { get; set; }
-        public int Summ { get; set; }
-            public virtual ICollection<Cost> Costs{ get; set; }
-            public int AcountId { get; set; }
-            public virtual Acount Acount { get; set; }
+
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public decimal Summ { get; set; }
+        public virtual ICollection<Cost> Costs { get; set; }
+        public int AcountId { get; set; }
+        public virtual Acount Acount { get; set; }
 
     }
 }
